Basic block replacement pass
